[Transport] Rename TransportBase.h to Transport.h
Also replace doublereal by double in Transport.h
